tests: queue tests fix
Before this patch, queue tests failed due to impossibility to use _queue_taken,
which was deprecated in tarantool/queue [1]. Switch to _queue_taken_2 fixed it.

Create access and some other actions were denied for test user.
Additional grants fixed this problem.

The test `TestUtube_Put` was affected by replacing timeout in the call
of `TakeTimeout()` with timeout of connection structure. For clarification
of this effect see annotation for function `TakeTimeout()` [2]. That was
fixed by making connection timeout longer than the timeout in the function call.

The step in testing.yml was uncommented for enabling queue tests in CI.

1. https://github.com/tarantool/queue/tree/7d05b6db5b0d596db4281d53b57811e156234a0d#fields-of-the-_queue_consumers-space
2. https://github.com/tarantool/go-tarantool/blob/master/queue/queue.go#L35

Fixes #115