sql: add minimal sql support
This patch adds the support of SQL in connector.
Added support of positional and named arguments.
Added ExecuteTyped() method for use with custom
packing/unpacking for a type.
Added all required constants to const.go for encoding SQL
in msgpack and decoding response.

Added SQL tests. Updated config.lua for creation the space for
using SQL in tests. Added the check of Tarantool version to skip SQL
tests if tarantool version < 2.0.0.

Changed id of the test spaces with id=512 and id=514, cause if using SQL
in tarantool there is no ability to set space id explicitly, so it gets
created with id=512 by default and conflicts with already existing
space with the same id.

Added new dependency in go.sum, go.mod for using assert package.

Added examples of using SQL queries in example_test.go
for compiling the future documentation from sources.

Added notes about the version since which Execute() is supported.

Closes #62