code health: decode schema with msgpack
When we decode the schema with msgpack it allows us to handle errors
better. For example, incorrect type conversion leads to a runtime
error. Now it will be an usual error that we can handle by our code.

Also it will help in migration to msgpack.v5, because an interface
decoding rules by default have changed in msgpack.v5.

Part of #124

Co-authored-by: Oleg Utkin <oleg_utkin@me.com>