msgpack: add optional msgpack.v5 support
The msgpack.v5 code located in msgpack_v5.go and
msgpack_v5_helper_test.go for tests. It is the same logic for
submodules.

An user can use msgpack.v5 with a build tag:
go_tarantool_msgpack_v5

The commit also unify typo conversions after decoding in tests. This
is necessary because  msgpack v2.9.2 decodes all numbers as uint[1]
or int[2], but msgpack.v5 decodes numbers as a MessagePack type[3].
There are additional differences when decoding types.

1. https://github.com/vmihailenco/msgpack/blob/23644a15054d8b9f8a9fca3e041f3419504b9c21/decode.go#L283
2. https://github.com/vmihailenco/msgpack/blob/23644a15054d8b9f8a9fca3e041f3419504b9c21/decode.go#L285
3. https://github.com/vmihailenco/msgpack/blob/233c977ae92b215a9aa7eb420bbe67da99f05ab6/decode.go#L410

Part of #124