api: add ConnectionHandler to the connection_pool
ConnectionHandler provides callbacks for components interested
in handling changes of connections in a ConnectionPool.

We have to take into account that user callbacks can take an
indefinite amount of time. The simplest solution is to process each
connection in a separate goroutine. This should not affect to
performance because most of the time these goroutines are blocked.

Closes #178