queue: encode UUID argument of identify() as string instead of binary
The identify() function expects the UUID argument to be a plain string
while the go connector encodes it in MsgPack as a binary blob (MP_BIN).
This works fine for now because Tarantool stores MP_BIN data in a string
when decoded to Lua but this behavior is going to change soon: we're
planning to introduce the new Lua type for binary data and update the
MsgPack decoder to store MP_BIN data in a varbianry object instead of
a plain string.

Let's prepare for that by converting the UUID data to a string before
encoding.

Needed for tarantool/tarantool#1629