test: separate SSL server and client check
Current SSL tests are as follows. We start a Tarantool server with
default helpers. "ok" tests are successful if everything had started,
"fail" tests are successful if ping check had failed (aka we failed to
connect). This is a dangerous approach, since "server had failed to
start" here is indistinguishable from "client cannot connect". Moreover,
because of it each tnt_fail test runs for 5 seconds
(10 retry attempts * 500 ms retry wait), which is frustrating. After
this patch, there is a separate check for a server start and for
a client success or fail.