test: remove core tarantool server SSL cases
"empty" and "key_crt_client" test cases do not provide SSL files to
a server started with SSL transport. In these cases server fails to
start, and tests ensures that server fails. It doesn't related to
go-tarantool connector testing in any way -- it's the test of
a tarantool binary. Since testing core tarantool is not the part of
go-tarantool project, this patch removes these cases.

The main motivation of this patch is the next commit in the patchset,
which separates check for server start and client success of fail.