crud: change timeout option type
`Timeout` is an option supported for all crud operations using
VShard API. In Lua, timeout has `number` type, so float values are
allowed. Float timeouts can be useful to set the timeout less than
a second. After this patch, `Timeout` will be an optional float64
instead of an optional int.

This is a breaking change.

CRUD allows to use negative timeouts. This approach does not make any
sense for an operation: for example, initial schema fetch will fail.
Since the module itself does not check for a negative value argument,
we do not forbid it here too.

Closes #342