api: add helper Dialer implementations
To disable SSL by default we want to transfer
`OpenSslDialer` to the go-openssl repository.
In order to do so, we need to minimize the amount of
copy-paste of the private functions.

`AuthDialer` is created as a dialer-wrapper, that
calls authentication methods.
`ProtoDialer` is created to receive and check the `ProtocolInfo`
in the created connection.
`GreetingDialer` is created to fill the `Greeting`
in the created connection.

Part of #301