crud: allow interface{} as values for *ManyRequest
It was a mistake to use `[]interface{}` or `[]msgpack.CustomEncoder`
as types for an array of tuples or an array of objects. Users were
unable to use slices of custom types as incoming values.

The patch now allows the use of `interface{}` as incoming values. It
makes it easier to use the API, but users need to be more careful.
Therefore, we have also added examples.

Closes #365