api: remove ssl
`OpenSslDialer` and all of its helper functions and tests
were rellocated to the `go-tlsdialer` [1] package (and renamed
to `OpenSSLDialer`). So now we can safely remove all the
copy-pasted code from `go-tarantool`.

This way, in order to use SSL, user should import the
`go-tlsdialer` package and call functions from there.

1. https://github.com/tarantool/go-tlsdialer/

Part of #301