api: remove Future.Err()
The method causes an improper error handling because it returns an error
only from a client side. Therefore, it is not enough to simply check
the error to find out that the request was not completed.

A user should check an error from `Future.Get()` or
`Future.GetTyped()`. In addition, the user can find out whether there
was an error without decoding the response body with
`Future.GetResponse()`, see `ExampleErrorNo`.