test: install crud from master
Currently `make deps`, which is used for tests, installs specified
version of `crud`. Such approach complicates integration with Tarantool:
when a patch in Tarantool breaks integration with both `crud` and
`go-tarantool` (because it uses `crud`), one needs to fix `crud`, then
release it and then bump it to the new version in makefile of
`go-tarantool`. Let's simplify this process - simply install `crud` from
its master for tests. Note that such approach has a drawback - any commit
in `crud` can break integration with `go-tarantool`.