tests: fix decode arrow check after update
The pull request changes [1] changes an error message on arrow
decoding. The patch replaces a check of error message to a check
of error code as it is a more stable check. However, the code is
changed too. So as a temporary solution we need to work with both
codes at the same time.

1. https://github.com/tarantool/tarantool/pull/10665

Part of #415