api: block Connect() on failure if Reconnect > 0
This patch makes Connect() retry connection attempts if opts.Reconnect
is greater than 0. The delay between connection attempts is
opts.Reconnect. If opts.MaxReconnects > 0 then the maximum number of
attempts is equal to it, otherwise the maximum number of attempts is
unlimited. Connect() now also blocks until a connection is established,
provided context is cancelled or the number of attempts is exhausted.

Closes #436